Quarto commit

+++ src/br/com/farmacia/modelo/dao/FiltroID.java
@@ -2,7 +2,7 @@ package br.com.farmacia.modelo.dao;
 
 public interface FiltroID<T> {
 
-	public T getBy(int id);
+	public T getBy(long id);
 
 	public T getBy(String id);
